[LINT] Use size_t for record and wait commands arguments
Fixes compiler warning:
src/commands.cpp: In member function ‘virtual bool Wait::ParseCommand(const char**, std::vector<Command*>&)’:
src/commands.cpp:422:32: warning: format ‘%u’ expects argument of type ‘unsigned int*’, but argument 3 has type ‘size_t* {aka long unsigned int*}’ [-Wformat=]
     sscanf(*cmds, "%u", &millis);
                         ~~~~~~~^

+++ src/channels.cpp
@@ -220,7 +220,7 @@ void TestChanAudio::FillPlaybackBuffer(char *buf, size_t len) {
         */
 }
 
-bool TestChanAudio::RecordAudioFile(PString &filename, bool append_file, bool stop_on_silence, int max_millisec) {
+bool TestChanAudio::RecordAudioFile(PString &filename, bool append_file, bool stop_on_silence, size_t max_millisec) {
     // std::cout << __func__ << std::endl;
     sync.Wait();
     /*
